requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.0
werkzeug>=3.0.0
anthropic>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
markdown>=3.5.0
pygments>=2.17.0
markupsafe>=2.1.0
